[refactor-#222] 스터디 화면 UI 컴포넌트 수정
